requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
openai>=1.12.0
requests==2.31.0
python-dotenv==1.0.0
sounddevice==0.4.6
soundfile==0.12.1
scipy>=1.14.0
opencv-python>=4.9.0.80
numpy>=1.26.3
pyttsx3==2.90
pynput==1.7.6
colorama==0.4.6
PyJWT>=2.8.0
livekit>=0.11.0
livekit-api>=0.5.0
deepface>=0.0.79
tf-keras>=2.15.0
